sandbox/chatplax/requirements.txt
openai
python-dotenv
plxscripting
